auth_service/requirements.txt
Django==4.2.30
djangorestframework==3.16.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.12
pika==1.3.2
requests==2.31.0
python-consul==1.1.0
